Rename finger printing to user agent reduction
