+++ packages/design-system/src/icons/index.tsx
@@ -34,8 +34,8 @@ export { default as BounceTrackingIcon } from './bounce-tracking.svg';
 export { default as BounceTrackingIconWhite } from './bounce-tracking-white.svg';
 export { default as TopicsIcon } from './topics.svg';
 export { default as TopicsIconWhite } from './topics-white.svg';
-export { default as FingerPrintingIcon } from './fingerprinting.svg';
-export { default as FingerPrintingIconWhite } from './fingerprinting-white.svg';
+export { default as UserAgentReductionIcon } from './user-agent-reduction.svg';
+export { default as UserAgentReductionIconWhite } from './user-agent-reduction-white.svg';
 export { default as InfoIcon } from './info.svg';
 export { default as SiteBoundariesIcon } from './site-boundaries.svg';
 export { default as SiteBoundariesIconWhite } from './site-boundaries-white.svg';
