+++ packages/extension/src/view/devtools/pages/antiCovertTracking/index.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 export { default as BounceTracking } from './bounceTracking';
-export { default as Fingerprinting } from './fingerprinting';
+export { default as UserAgentReduction } from './userAgentReduction';
 export { default as IPProtection } from './ipProtection';
 export { default as PrivateStateTokens } from './privateStateTokens';
 export { default as AntiCovertTracking } from './antiCovertTracking';
