+++ packages/extension/src/view/devtools/pages/antiCovertTracking/userAgentReduction/index.tsx
@@ -20,7 +20,7 @@ import React from 'react';
 import { LandingPage, PSInfoKey } from '@google-psat/design-system';
 import { I18n } from '@google-psat/i18n';
 
-const Fingerprinting = () => {
+const UserAgentReduction = () => {
   return (
     <div data-testid="fingerprinting-content" className="h-full w-full">
       <LandingPage
@@ -32,4 +32,4 @@ const Fingerprinting = () => {
   );
 };
 
-export default Fingerprinting;
+export default UserAgentReduction;
