+++ packages/extension/src/view/devtools/tabs.ts
@@ -27,8 +27,8 @@ import {
   AttributionIcon,
   BounceTrackingIcon,
   BounceTrackingIconWhite,
-  FingerPrintingIcon,
-  FingerPrintingIconWhite,
+  UserAgentReductionIcon,
+  UserAgentReductionIconWhite,
   TopicsIcon,
   PrivacySandboxIcon,
   PrivacySandboxIconWhite,
@@ -69,7 +69,7 @@ import {
   Topics,
   AttributionReporting,
   BounceTracking,
-  Fingerprinting,
+  UserAgentReduction,
   PrivacySandbox,
   Wiki,
   Settings,
@@ -307,7 +307,7 @@ const TABS: SidebarItems = {
         },
       },
       [SIDEBAR_ITEMS_KEYS.ANTI_COVERT_TRACKING]: {
-        title: () => I18n.getMessage('trackingProtection'),
+        title: () => 'Privacy Protection',
         panel: {
           Element: AntiCovertTracking,
         },
@@ -350,22 +350,34 @@ const TABS: SidebarItems = {
             },
             icon: {
               Element: BounceTrackingIcon,
+              props: {
+                className: 'relative top-[2px]',
+              },
             },
             selectedIcon: {
               Element: BounceTrackingIconWhite,
+              props: {
+                className: 'relative top-[2px]',
+              },
             },
             children: {},
           },
           [SIDEBAR_ITEMS_KEYS.FINGERPRINTING]: {
-            title: () => I18n.getMessage('fingerprinting'),
+            title: () => 'User Agent Reduction',
             panel: {
-              Element: Fingerprinting,
+              Element: UserAgentReduction,
             },
             icon: {
-              Element: FingerPrintingIcon,
+              Element: UserAgentReductionIcon,
+              props: {
+                className: 'relative top-[2px]',
+              },
             },
             selectedIcon: {
-              Element: FingerPrintingIconWhite,
+              Element: UserAgentReductionIconWhite,
+              props: {
+                className: 'relative top-[2px]',
+              },
             },
             children: {},
           },
